src/requirements.txt
aniso8601==1.2.1
boto3==1.7.70
botocore==1.10.70
certifi==2017.4.17
chardet==3.0.4
click==6.7
docutils==0.14
Flask==0.12.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.6
hvac==0.2.17
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
PyJWT==1.4.2
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.13
requests==2.18.1
s3transfer==0.1.13
six==1.10.0
urllib3==1.21.1
Werkzeug==0.12.2
